delete client method updated

+++ apitest/views.py
@@ -932,6 +932,18 @@ class ClientDetailsAPI(ModelViewSet):
     def destroy(self, request, *args, **kwargs):
         try:
             instance = self.get_object()
+
+            # Check if there are associated records in Estimatedetails table
+            if Estimatedetails.objects.filter(clientid=instance.clientid).exists():
+                error_message = "Client has estimates so it cant be deleted"
+                error_response = {
+                    "status": "success",
+                    "code": status.HTTP_400_BAD_REQUEST,
+                    "message": error_message,
+                }
+                return Response(error_response, status=status.HTTP_400_BAD_REQUEST)
+
+            # If no associated records found, proceed with deletion
             instance.delete()
             api_response = {
                 "status": "success",
